+++ numba/cuda/initialize.py
@@ -3,7 +3,7 @@ def initialize_all():
     import numba.cuda.models  # noqa: F401
 
     from numba import cuda
-    from numba.cuda.compiler import Dispatcher
+    from numba.cuda.dispatcher import Dispatcher
     from numba.core.target_extension import (target_registry,
                                              dispatcher_registry,
                                              jit_registry)
